Fix Telegram provider authentication flow
