+++ packages/jinn/src/connectors/telegram/__tests__/connector.test.ts
@@ -1,49 +1,18 @@
-import { describe, it, expect, vi, beforeEach, afterEach } from "vitest";
+import { describe, it, expect, vi, beforeEach } from "vitest";
 import type { IncomingMessage, Session, Target } from "../../../shared/types.js";
 
 // Mock node-telegram-bot-api before importing connector
 const mockSendMessage = vi.fn().mockResolvedValue({ message_id: 1 });
 const mockEditMessageText = vi.fn().mockResolvedValue(true);
-const mockDeleteMessage = vi.fn().mockResolvedValue(true);
 const mockGetMe = vi.fn().mockResolvedValue({ id: 999, username: "test_bot" });
 const mockStartPolling = vi.fn();
 const mockStopPolling = vi.fn().mockResolvedValue(undefined);
 const mockOn = vi.fn();
-const mockPtyWrite = vi.fn();
-const mockPtyKill = vi.fn();
-const mockPtyOnData = vi.fn();
-const mockPtyOnExit = vi.fn();
-const mockPtySpawn = vi.fn(() => ({
-  write: mockPtyWrite,
-  kill: mockPtyKill,
-  onData: mockPtyOnData,
-  onExit: mockPtyOnExit,
-}));
-let mockClaudeLoggedIn = false;
-let mockCodexExitCode = 1;
-let mockCodexHang = false;
-const mockExecFile = vi.fn((_file, _args, _options, callback) => {
-  callback(null, JSON.stringify({ loggedIn: mockClaudeLoggedIn }));
-});
-const mockChildKill = vi.fn();
-const mockSpawnRemoveListener = vi.fn();
-const mockSpawnOnce = vi.fn(function (this: any, event: string, handler: (code?: number) => void) {
-  if (event === "exit" && !mockCodexHang) {
-    queueMicrotask(() => handler(mockCodexExitCode));
-  }
-  return this;
-});
-const mockSpawn = vi.fn(() => ({
-  kill: mockChildKill,
-  once: mockSpawnOnce,
-  removeListener: mockSpawnRemoveListener,
-}));
 
 vi.mock("node-telegram-bot-api", () => {
   const MockBot = vi.fn(function (this: any) {
     this.sendMessage = mockSendMessage;
     this.editMessageText = mockEditMessageText;
-    this.deleteMessage = mockDeleteMessage;
     this.getMe = mockGetMe;
     this.startPolling = mockStartPolling;
     this.stopPolling = mockStopPolling;
@@ -51,15 +20,6 @@ vi.mock("node-telegram-bot-api", () => {
   });
   return { default: MockBot };
 });
-
-vi.mock("node-pty", () => ({
-  spawn: mockPtySpawn,
-}));
-
-vi.mock("node:child_process", () => ({
-  execFile: mockExecFile,
-  spawn: mockSpawn,
-}));
 
 vi.mock("../../../shared/logger.js", () => ({
   logger: {
@@ -79,16 +39,9 @@ describe("TelegramConnector", () => {
 
   beforeEach(() => {
     vi.clearAllMocks();
-    mockClaudeLoggedIn = false;
-    mockCodexExitCode = 1;
-    mockCodexHang = false;
     connector = new TelegramConnector({
       botToken: "123456:ABC-DEF",
     });
-  });
-
-  afterEach(() => {
-    vi.useRealTimers();
   });
 
   describe("constructor", () => {
@@ -145,324 +98,6 @@ describe("TelegramConnector", () => {
   });
 
   describe("onMessage", () => {
-    it("preserves default-disabled auth text routing", async () => {
-      const handler = vi.fn();
-      connector.onMessage(handler);
-      await connector.start();
-
-      const messageCallback = mockOn.mock.calls.find(
-        (call) => call[0] === "message",
-      )?.[1];
-      await messageCallback({
-        message_id: 42,
-        chat: { id: 12345, type: "private" as const },
-        from: { id: 67890, username: "testuser", first_name: "Test", is_bot: false },
-        date: Math.floor(Date.now() / 1000) + 10,
-        text: "/auth status",
-      });
-
-      expect(handler).toHaveBeenCalledOnce();
-      expect(handler.mock.calls[0][0].text).toBe("/auth status");
-      expect(mockSendMessage).not.toHaveBeenCalled();
-      expect(mockPtySpawn).not.toHaveBeenCalled();
-    });
-
-    it("handles owner auth status without a normal handler", async () => {
-      const authConnector = new TelegramConnector({
-        botToken: "123456:ABC-DEF",
-        allowFrom: [67890],
-        telegramAuth: {
-          enabled: true,
-          ownerUserIds: [67890],
-        },
-      });
-      await authConnector.start();
-
-      const messageCallback = mockOn.mock.calls.find(
-        (call) => call[0] === "message",
-      )?.[1];
-      await messageCallback({
-        message_id: 42,
-        chat: { id: 12345, type: "private" as const },
-        from: { id: 67890, username: "owner", first_name: "Owner", is_bot: false },
-        date: Math.floor(Date.now() / 1000) + 10,
-        text: "/auth status",
-      });
-
-      expect(mockSendMessage).toHaveBeenCalledWith(
-        "12345",
-        [
-          "Claude is not authenticated. Use /auth claude to sign in.",
-          "Codex is not authenticated. Use /auth codex to sign in.",
-        ].join("\n"),
-        { parse_mode: "Markdown" },
-      );
-      expect(mockPtySpawn).not.toHaveBeenCalled();
-    });
-
-    it("reports authenticated provider status without CLI output", async () => {
-      mockClaudeLoggedIn = true;
-      mockCodexExitCode = 0;
-      const authConnector = new TelegramConnector({
-        botToken: "123456:ABC-DEF",
-        allowFrom: [67890],
-        telegramAuth: {
-          enabled: true,
-          ownerUserIds: [67890],
-        },
-      });
-      const handler = vi.fn();
-      authConnector.onMessage(handler);
-      await authConnector.start();
-
-      const messageCallback = mockOn.mock.calls.find(
-        (call) => call[0] === "message",
-      )?.[1];
-      await messageCallback({
-        message_id: 42,
-        chat: { id: 12345, type: "private" as const },
-        from: { id: 67890, username: "owner", first_name: "Owner", is_bot: false },
-        date: Math.floor(Date.now() / 1000) + 10,
-        text: "/auth status",
-      });
-
-      expect(handler).not.toHaveBeenCalled();
-      expect(mockSendMessage).toHaveBeenCalledWith(
-        "12345",
-        ["Claude is authenticated.", "Codex is authenticated."].join("\n"),
-        { parse_mode: "Markdown" },
-      );
-      expect(mockExecFile).toHaveBeenCalledWith(
-        "claude",
-        ["auth", "status", "--json"],
-        expect.objectContaining({
-          env: expect.objectContaining({
-            HOME: "/home/node",
-            CLAUDE_CONFIG_DIR: "/home/node/.claude",
-            CODEX_HOME: "/home/node/.codex",
-          }),
-        }),
-        expect.any(Function),
-      );
-      expect(mockSpawn).toHaveBeenCalledWith(
-        "codex",
-        ["login", "status"],
-        expect.objectContaining({
-          env: expect.objectContaining({
-            HOME: "/home/node",
-            CLAUDE_CONFIG_DIR: "/home/node/.claude",
-            CODEX_HOME: "/home/node/.codex",
-          }),
-          stdio: "ignore",
-        }),
-      );
-    });
-
-    it("terminates timed-out Codex status and escalates to SIGKILL", async () => {
-      vi.useFakeTimers();
-      mockClaudeLoggedIn = true;
-      mockCodexHang = true;
-      const authConnector = new TelegramConnector({
-        botToken: "123456:ABC-DEF",
-        allowFrom: [67890],
-        telegramAuth: {
-          enabled: true,
-          ownerUserIds: [67890],
-        },
-      });
-      await authConnector.start();
-
-      const messageCallback = mockOn.mock.calls.find(
-        (call) => call[0] === "message",
-      )?.[1];
-      const statusPromise = messageCallback({
-        message_id: 42,
-        chat: { id: 12345, type: "private" as const },
-        from: { id: 67890, username: "owner", first_name: "Owner", is_bot: false },
-        date: Math.floor(Date.now() / 1000) + 10,
-        text: "/auth status",
-      });
-
-      await vi.advanceTimersByTimeAsync(15_000);
-      expect(mockChildKill).toHaveBeenCalledWith("SIGTERM");
-      await vi.advanceTimersByTimeAsync(2_000);
-      expect(mockChildKill).toHaveBeenCalledWith("SIGKILL");
-      await vi.advanceTimersByTimeAsync(2_000);
-      await statusPromise;
-
-      expect(mockSpawnRemoveListener).toHaveBeenCalledWith("error", expect.any(Function));
-      expect(mockSpawnRemoveListener).toHaveBeenCalledWith("exit", expect.any(Function));
-      expect(mockSpawnRemoveListener).toHaveBeenCalledWith("close", expect.any(Function));
-      expect(mockSendMessage).toHaveBeenCalledWith(
-        "12345",
-        [
-          "Claude is authenticated.",
-          "Codex is not authenticated. Use /auth codex to sign in.",
-        ].join("\n"),
-        { parse_mode: "Markdown" },
-      );
-    });
-
-    it("starts owner auth without calling the normal handler", async () => {
-      const authConnector = new TelegramConnector({
-        botToken: "123456:ABC-DEF",
-        allowFrom: [67890],
-        telegramAuth: {
-          enabled: true,
-          ownerUserIds: [67890],
-        },
-      });
-      const handler = vi.fn();
-      authConnector.onMessage(handler);
-      await authConnector.start();
-
-      const messageCallback = mockOn.mock.calls.find(
-        (call) => call[0] === "message",
-      )?.[1];
-      await messageCallback({
-        message_id: 43,
-        chat: { id: 12345, type: "private" as const },
-        from: { id: 67890, username: "owner", first_name: "Owner", is_bot: false },
-        date: Math.floor(Date.now() / 1000) + 10,
-        text: "/auth claude",
-      });
-
-      expect(handler).not.toHaveBeenCalled();
-      expect(mockPtySpawn).toHaveBeenCalledWith(
-        "claude",
-        ["auth", "login", "--claudeai"],
-        expect.objectContaining({
-          cwd: "/home/node",
-          env: expect.objectContaining({
-            HOME: "/home/node",
-            CLAUDE_CONFIG_DIR: "/home/node/.claude",
-            CODEX_HOME: "/home/node/.codex",
-          }),
-        }),
-      );
-    });
-
-    it("deletes owner auth input best-effort without calling the normal handler", async () => {
-      const authConnector = new TelegramConnector({
-        botToken: "123456:ABC-DEF",
-        allowFrom: [67890],
-        telegramAuth: {
-          enabled: true,
-          ownerUserIds: [67890],
-        },
-      });
-      const handler = vi.fn();
-      authConnector.onMessage(handler);
-      await authConnector.start();
-
-      const messageCallback = mockOn.mock.calls.find(
-        (call) => call[0] === "message",
-      )?.[1];
-      await messageCallback({
-        message_id: 44,
-        chat: { id: 12345, type: "private" as const },
-        from: { id: 67890, username: "owner", first_name: "Owner", is_bot: false },
-        date: Math.floor(Date.now() / 1000) + 10,
-        text: "/auth input AB12-CD34",
-      });
-
-      expect(handler).not.toHaveBeenCalled();
-      expect(mockDeleteMessage).toHaveBeenCalledWith("12345", 44);
-    });
-
-    it("ignores non-owner auth text after allowFrom without normal routing", async () => {
-      const authConnector = new TelegramConnector({
-        botToken: "123456:ABC-DEF",
-        allowFrom: [67890, 99999],
-        telegramAuth: {
-          enabled: true,
-          ownerUserIds: [67890],
-        },
-      });
-      const handler = vi.fn();
-      authConnector.onMessage(handler);
-      await authConnector.start();
-
-      const messageCallback = mockOn.mock.calls.find(
-        (call) => call[0] === "message",
-      )?.[1];
-      await messageCallback({
-        message_id: 45,
-        chat: { id: 12345, type: "private" as const },
-        from: { id: 99999, username: "stranger", first_name: "Stranger", is_bot: false },
-        date: Math.floor(Date.now() / 1000) + 10,
-        text: "/auth status",
-      });
-
-      expect(handler).not.toHaveBeenCalled();
-      expect(mockSendMessage).not.toHaveBeenCalled();
-      expect(mockPtySpawn).not.toHaveBeenCalled();
-    });
-
-    it("rejects owner auth commands in groups without normal routing", async () => {
-      const authConnector = new TelegramConnector({
-        botToken: "123456:ABC-DEF",
-        allowFrom: [67890],
-        telegramAuth: {
-          enabled: true,
-          ownerUserIds: [67890],
-        },
-      });
-      const handler = vi.fn();
-      authConnector.onMessage(handler);
-      await authConnector.start();
-
-      const messageCallback = mockOn.mock.calls.find(
-        (call) => call[0] === "message",
-      )?.[1];
-      await messageCallback({
-        message_id: 46,
-        chat: { id: -10012345, type: "group" as const },
-        from: { id: 67890, username: "owner", first_name: "Owner", is_bot: false },
-        date: Math.floor(Date.now() / 1000) + 10,
-        text: "/auth status",
-      });
-
-      expect(handler).not.toHaveBeenCalled();
-      expect(mockSendMessage).toHaveBeenCalledWith(
-        "-10012345",
-        "Authentication commands are available only in a private chat.",
-        { parse_mode: "Markdown" },
-      );
-    });
-
-    it("routes ordinary owner messages unchanged when auth is enabled", async () => {
-      const authConnector = new TelegramConnector({
-        botToken: "123456:ABC-DEF",
-        allowFrom: [67890],
-        telegramAuth: {
-          enabled: true,
-          ownerUserIds: [67890],
-        },
-      });
-      const handler = vi.fn();
-      authConnector.onMessage(handler);
-      await authConnector.start();
-
-      const messageCallback = mockOn.mock.calls.find(
-        (call) => call[0] === "message",
-      )?.[1];
-      await messageCallback({
-        message_id: 47,
-        chat: { id: 12345, type: "private" as const },
-        from: { id: 67890, username: "owner", first_name: "Owner", is_bot: false },
-        date: Math.floor(Date.now() / 1000) + 10,
-        text: "Hello bot!",
-      });
-
-      expect(handler).toHaveBeenCalledOnce();
-      const msg: IncomingMessage = handler.mock.calls[0][0];
-      expect(msg.text).toBe("Hello bot!");
-      expect(msg.userId).toBe("67890");
-      expect(msg.channel).toBe("12345");
-      expect(mockSendMessage).not.toHaveBeenCalled();
-    });
-
     it("stamps and replies through a named connector instance id", async () => {
       const named = new TelegramConnector({
         id: "telegram-support",
@@ -660,140 +295,6 @@ describe("TelegramConnector", () => {
     it("rejects when the plain-text retry also fails", async () => {
       mockSendMessage.mockRejectedValueOnce(new Error("can't parse entities")).mockRejectedValueOnce(new Error("chat not found"));
       await expect(connector.sendMessage(target, "**bad markdown")).rejects.toThrow("chat not found");
-    });
-  });
-
-  describe("authentication failures", () => {
-    it("offers the owner login commands after a provider authentication failure", async () => {
-      const authConnector = new TelegramConnector({
-        botToken: "123456:ABC-DEF",
-        allowFrom: [67890],
-        telegramAuth: {
-          enabled: true,
-          ownerUserIds: [67890],
-        },
-      });
-
-      await authConnector.replyMessage(
-        {
-          channel: "12345",
-          replyContext: {
-            chatId: "12345",
-            messageId: 42,
-            chatType: "private",
-            userId: 67890,
-          },
-        },
-        "⛔ Interactive turn failed: authentication_failed",
-      );
-
-      expect(mockSendMessage).toHaveBeenCalledWith(
-        "12345",
-        expect.stringContaining("use /auth claude or /auth codex to sign in."),
-        expect.objectContaining({
-          parse_mode: "Markdown",
-          reply_parameters: { message_id: 42 },
-        }),
-      );
-    });
-
-    it("does not add login prompts to unrelated provider errors", async () => {
-      const authConnector = new TelegramConnector({
-        botToken: "123456:ABC-DEF",
-        allowFrom: [67890],
-        telegramAuth: {
-          enabled: true,
-          ownerUserIds: [67890],
-        },
-      });
-
-      await authConnector.replyMessage(
-        {
-          channel: "12345",
-          replyContext: {
-            chatId: "12345",
-            messageId: 42,
-            chatType: "private",
-            userId: 67890,
-          },
-        },
-        "⛔ unrelated MCP: browser is not logged in",
-      );
-
-      expect(mockSendMessage).toHaveBeenCalledWith(
-        "12345",
-        "⛔ unrelated MCP: browser is not logged in",
-        {
-          parse_mode: "Markdown",
-          reply_parameters: { message_id: 42 },
-        },
-      );
-    });
-
-    it("does not offer login outside the owner private chat", async () => {
-      const authConnector = new TelegramConnector({
-        botToken: "123456:ABC-DEF",
-        allowFrom: [67890],
-        telegramAuth: {
-          enabled: true,
-          ownerUserIds: [67890],
-        },
-      });
-
-      await authConnector.replyMessage(
-        {
-          channel: "-100999",
-          replyContext: {
-            chatId: "-100999",
-            messageId: 42,
-            chatType: "group",
-            userId: 67890,
-          },
-        },
-        "⛔ Interactive turn failed: authentication_failed",
-      );
-
-      expect(mockSendMessage).toHaveBeenCalledWith(
-        "-100999",
-        "⛔ Interactive turn failed: authentication_failed",
-        {
-          parse_mode: "Markdown",
-          reply_parameters: { message_id: 42 },
-        },
-      );
-    });
-
-    it("does not offer login to a non-owner private chat", async () => {
-      const authConnector = new TelegramConnector({
-        botToken: "123456:ABC-DEF",
-        allowFrom: [11111],
-        telegramAuth: {
-          enabled: true,
-          ownerUserIds: [67890],
-        },
-      });
-
-      await authConnector.replyMessage(
-        {
-          channel: "12345",
-          replyContext: {
-            chatId: "12345",
-            messageId: 42,
-            chatType: "private",
-            userId: 11111,
-          },
-        },
-        "⛔ Interactive turn failed: authentication_failed",
-      );
-
-      expect(mockSendMessage).toHaveBeenCalledWith(
-        "12345",
-        "⛔ Interactive turn failed: authentication_failed",
-        {
-          parse_mode: "Markdown",
-          reply_parameters: { message_id: 42 },
-        },
-      );
     });
   });
